requirements.txt
# TMDD - Threat Modeling Driven Development
# Required dependencies

PyYAML>=6.0

